test (ontology util): move test
